Partially handle EAX1.0 properties
This is kind of dodgy since it tries to use the EAX2 environment presets while
mapping Volume to the effect/room gain (converting between linear gain and mB
as necessary), passing the decay time through as-is, and pretending to handle
dampening (the value is tracked, but ignored). Similarly, the EAX1 buffer
property maps the reverb mix to the room path gain, converting between linear
gain and mB.

This will hopefully be enough to satisfy apps that use EAX 1.0 while being
relatively close to the intended sound.