Reference DeviceShare members through the share itself
Rather than pointers to individual members of the DeviceShare, just use the
pointer to the DeviceShare itself. There's no extra indirection since it's just
replacing one pointer with another, with an added offset.

This doesn't apply to members that are copied, such as the ALCcontext handle or
the auxiliary effect slot ID.