Avoid caching the thread-local context
There are issues with the cached value becoming stale, resulting in the context
not being set current after being destroed and recreated at the same address.

Would be nice to have an "explicit" API for OpenAL calls, so the context can be
specified explicitly to each call instead of having to make it current.